Add common database provider
